+++ src/main/java/me/pajic/enchantmentdisabler/Main.java
@@ -3,7 +3,10 @@ package me.pajic.enchantmentdisabler;
 import me.fzzyhmstrs.fzzy_config.api.ConfigApiJava;
 import me.pajic.enchantmentdisabler.config.ModConfig;
 import me.pajic.enchantmentdisabler.mixson.ResourceModifications;
+import me.pajic.enchantmentdisabler.util.ModUtil;
 import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.event.registry.DynamicRegistrySetupCallback;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
 
 public class Main implements ModInitializer {
@@ -14,5 +17,9 @@ public class Main implements ModInitializer {
     @Override
     public void onInitialize() {
         ResourceModifications.init();
+        DynamicRegistrySetupCallback.EVENT.register(registryView -> registryView.registerEntryAdded(
+                Registries.ENCHANTMENT,
+                (rawId, id, object) -> ModUtil.enchantmentSuggestions.add(id)
+        ));
     }
 }